[mojo-lang] Make `@value` generate trait conformances early (#33906)
This PR is a bit ugly but it makes it so that `@value` will generate the
trait conformances for `Copyable` and `Movable` early. The trait is
added during signature resolution, which is needed because it is part of
the signature of the type, and in body resolution, before any of the
field types are resolved, the `__copyinit__` and `__moveinit__` stubs
are generated. This ensures that if there are recursive references to
the struct type during resolution of the field types, the trait
conformances are satisfied.

This PR also includes some code refactoring throughout StructEmitter.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: e6a84e152f4f0793b3bff3757f9258c220b61343