[Support][Cache] Check for available space before writing file. (#32906)
There is a chance we will do partial writes when the disk space is not enough, and since sometimes this is done under lock file we might even hang. So add a sanity check to exit early for this case.

MODULAR_ORIG_COMMIT_REV_ID: bb2952817dacf1d93e447871afe0fef625dae4f4