[Mojo-stdlib] [PythonInterop] Make __iter__ act more like Python (#33874)
Previously we had `__iter__` return an empty iterator if the object was
not an iterator or a sequence. However, there are more iterable objects
in Python than iterators or sequences, so remove this check. Also bubble
up the exception from Python if the object is not iterable, rather than
silently returning an empty iterator.

While we're at it, `__iter__` does not need a mutable reference to
`self`, so change that too to allow more ergonomic code like `for value
in my_dict.values():` without having to introduce an extra `var
dict_values = my_dict.values()` before iterating as was previously
required.

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>

modular-orig-commit: c9cbcd8ca1a888cb4de6d69983f841d86139b2e1