[Stdlib] Make abort able to take a return type (#34026)
This allows one to use abort to return a null return type. Note that the
code after abort is not reachable so it's ok to generate code that uses
undefined behavior (in this case a null deref).

So code which uses abort -> return is switch to use `return abort()`,
but the main user of this is going to be in
[Internal Link] where we want to abort
with an error but also return a valid return value.

modular-orig-commit: fc0ea0994c58d4535948e23f5c17ae8f8168d2cb