[mojo-lang] Enable implicit conformance of nominal types to traits (#33999)
This PR enables implicit conformance of nominal types to traits. This is
performed by changing `checkNominalTypeConformance` to first check for
an explicit conformance, and if this fails, it attempts to verify
conformance of the trait as if it explicitly conformed. If this
succeeds, it will tag the struct with the conformance so that this isn't
computed multiple times.

MODULAR_ORIG_COMMIT_REV_ID: c9824e7dbf5eff720e1696a23ff20ffd132abe3c