[mojo-lang] Print function types for generated functions (#34030)
When emitting a "candidate declared here" note for a generated function,
the user can't actually see the signature of the generated function, so
print it for the user to see.

Also, the convention was being printed after the argument name, and
argument
name used a `=` instead of `:` for function types.

MODULAR_ORIG_COMMIT_REV_ID: b92f1953b7f493ba420f01c182886688da73adf8