[Support][DebugInfo] Debug Value to Declare handle undefs (#33785)
Allow undefs to kill a variable.

Once a variable is declared with a llvm.dbg.declare, it is always
considered to live as long as the scope. This means to shorten the
lifetime, we need to use llvm.dbg.value instead, which allows encoding a
"killed" value with an undef operand, and also allows multiple
llvm.dbg.values to together describe the locations of a value across a
scope.

The latest implementation is this:
- [Existing] If a variable is only ever described by one dbg.value, turn
it into a stack allocation and use dbg.declare instead.
- [Existing] If a variable is described by multiple non-undef dbg.value,
don't transform anything. Keep using dbg.value.
- [New] If a variable is described by one non-undef dbg.value, and
multiple undef dbg.value, turn it into a stack allocation and wrap the
non-undef dbg.value with a deref DIExpr operator. Keep the undef
dbg.values the same.

This new policy has the benefit of both worlds: By using dbg.value, it
allows the variable to have a shorter lifetime than the entire scope. By
describing the location of the variable with a stack allocation, it
makes the debuginfo likely to survive the llvm backend (unlike raw ssa
values, which are very prone to losing debuginfo during instruction
selection). As can be seen, we still preserve debuginfo for all
variables in our lldb tests, which is not the case if we didn't allocate
variables to the stack.

I also took the chance to refactor the whole thing as it was getting
complicated all in one function.

MODULAR_ORIG_COMMIT_REV_ID: 6cad3bcabae6c67ac875761b7904ec8c7ad0a8c6