[Kernels] Pack the A matrix with i8mm and batched_matmul (#34015)
With i8mm which Graviton3 and the M3 processor have when doing int8
matmul we pack the A matrix to improve performance. This was being done
for matmul but not batched_matmul. This PR does it for batched_matmul as
well.

Fixes #32400

---------

Co-authored-by: Tracy Sharpe <99041446+raiseirql@users.noreply.github.com>
LINALG_ORIG_REV_ID: 5017e601769da2b573389cb88f3a48d1bc1b7940