[LSP] Add a global rerun for python tests (#34229)
It's a fact that the flakes we've been having are because of the parsing
crashing. Even when the tests run serially, the parser might crash, so
it could be related to some other mojo tests running in parallel. It
would be interesting to make these tests run serially among all mojo
tests. In any case, these tests run so fast that I'm adding a retry.
I still don't want to disable all of them because a few times they have
caught actual issues. For example, I helped Steffi this week with a
legit LSP crash caused due to her changes. So, until we figure out why
the parser crashes, I prefer to add as many retries as needed.

MODULAR_ORIG_COMMIT_REV_ID: 66a3ac9a520d818405429392b1daf54335357eb2