[mojo-lang][CheckLifetimes] Fix handling of break/continue/raise in if's (#34335)
The uninit checker wasn't handling the merge point of these terminators
correctly because it wasn't treating them the same way as
return/unreachable.  The right behavior (as far as the uninit checker
pass is concerned) is for any values to be fully defined after the
terminator as is done in checkTerminatorOp.

This fixes #34320.  Many thanks to Jeff for noticing the issue, and
even more for writing the testcase, which is the majority of the work
here.

MODULAR_ORIG_COMMIT_REV_ID: a8ef7baedec30107a97c21dc2c863a3999f5af05