[mojo-lang][CheckLifetimes] Improve temp elision optimization. (#34339)
This improves the temporary vardecl copy elision optimization to
handle `lit.ref.immut` and to not use n^2 scanning.  This allows
it to be more aggressive and unnerf's it a bit from the lifetime
changes I put in.

This isn't enough to fix #34138 but does include a testcase for it.

MODULAR_ORIG_COMMIT_REV_ID: e90199e9331f241a7d25a6f1dbce5aed6e30f1a2