[ONNX] Convert sig before moving monnx.graph block (#34334)
Convert the types of the `monnx.graph` signature prior to moving it to
the newly created `mo.graph`.
The existing ordering causes a failed legalization when we move the
block to the new op prior to type conversion.
This seems related to the order in which the rewriter pushes
`IRRewrite`s to the rewrite stack during dialect conversion.
Recent major refactoring in this area probably triggered this.
See for example LLVM [PR
81237](https://github.com/llvm/llvm-project/pull/81237).

As a drive by remove an unused variable in ExecuteOp.cpp causing a build
warning.

Testing:
- Manually, update to the latest LLVM (#34294).
  Then run `build check-onnx`.
  Prior to this change, all `monnx.graph` conversions fail to legalize.
  With this change they are fine.

MODULAR_ORIG_COMMIT_REV_ID: 76b71abf4580ce6d6a3044e7e474b93f90441eb6