[mojo-lang] Refactor handling of parameter packs. (#34348)
The implementation of parameter packs is old and from another time.
It converted an unpack of an type list into a PackType in very
generic code (ExprNode) even though PackType is an
argument-list-specific representation.  The "*" in an argument type
unpack isn't actually a general unpack, it is a designator that the
type expression is actually a type list.

This rearranges things to keep the logic for handling this separate.
This will help us implement general purpose unpacking logic sometime
in the future without interfering with the argument list specific
logic that we need now.

This doesn't actually use the fancy new rep, it just moves things
closer to supporting that.

MODULAR_ORIG_COMMIT_REV_ID: 73f2d5967e6caa27bd7604218273686a9038ce4f