[mojo-lang] Simplify logic for parameter unpacking. ~NFC (#34357)
This significantly simplifies the representation and implementation
of parameter unpacking. It was implemented in a very general way,
even though it is solving a specific problem.  The Mojo parser
cannot do unpacking of parameters at comptime - because general
pack values are only known at elaboration time they cannot be used
for parser-time binding.

This implication of this is that `*_` is not a generalized unpack
expression that can occur anywhere, it is magic syntax of a
parameter list itself.

This patch carves back UnpackAttr to be more specific about what it
is used for, clarifying the code and allowing simplifications.  IMO,
we should go further and eliminate the expression form and attribute
entirely, but that deserves more discussion.

This is NFC except for changing the order of emission of a
diagnostic so one testcase with two errors on the same line changes
the error it emits.

MODULAR_ORIG_COMMIT_REV_ID: 2aafaabb0ec5f61018593570d1c72a92ce46385e