[LSP] Enable crash recovery conditionally (#34417)
This enables crash recovery only if a certain env var is passed. I'm
going to run this on the CI many times and make sure this isn't flaky
anymore.

The crashes we've been having had this stack trace, which makes me think
that there's a race condition when multiple threads try to use the
`RunSafely` functionality, even if not at the same time. That
`__cxa_deleted_virtual` seems like we are calling a function that is
marked as deleted in the vtable? Super weird.

Another option is to try `RunSafelyOnThread`, but at this point I'd
rather nuke that functionality in the CI.

```
re virtual method called
1: terminate called without an active exception
1: Please submit a bug report to [Internal link] and include the crash backtrace along with all the relevant source codes with the contents they had at crash time.
1:  #0 0x0000aaaab8d88db8 llvm::sys::PrintStackTrace(llvm::raw_ostream&, int) Signals.cpp:0:0
1:  #1 0x0000aaaab8d86b6c llvm::sys::RunSignalHandlers() Signals.cpp:0:0
1:  #2 0x0000aaaab8d880b4 llvm::sys::CleanupOnSignal(unsigned long) Signals.cpp:0:0
1:  #3 0x0000aaaab8d40d04 CrashRecoverySignalHandler(int) CrashRecoveryContext.cpp:0:0
1:  #4 0x0000ffffb4212850 (linux-vdso.so.1+0x850)
1:  #5 0x0000ffffb3cc7d78 raise (/lib/aarch64-linux-gnu/libc.so.6+0x33d78)
1:  #6 0x0000ffffb3cb4aac abort (/lib/aarch64-linux-gnu/libc.so.6+0x20aac)
1:  #7 0x0000ffffb3ecc82c __gnu_cxx::__verbose_terminate_handler() (/lib/aarch64-linux-gnu/libstdc++.so.6+0xa182c)
1:  #8 0x0000ffffb3eca20c (/lib/aarch64-linux-gnu/libstdc++.so.6+0x9f20c)
1:  #9 0x0000ffffb3eca270 (/lib/aarch64-linux-gnu/libstdc++.so.6+0x9f270)
1: #10 0x0000ffffb3ecb290 __cxa_deleted_virtual (/lib/aarch64-linux-gnu/libstdc++.so.6+0xa0290)
1: #11 0x0000aaaab8d40b3c llvm::CrashRecoveryContext::RunSafely(llvm::function_ref<void ()>) CrashRecoveryContext.cpp:0:0
1: #12 0x0000aaaab8cd9b60 M::Mojo::LSP::MojoDocument::parseDocument() /__w/modular/modular/KGEN/tools/mojo-lsp-server/MojoServer.cpp:640:7
```

MODULAR_ORIG_COMMIT_REV_ID: 93a434bf41e94638b4d150392e69b632a185ed17