[mojo-lang] Refactor `!lit.ref.pack` representation a bit, NFC (#34618)
This moves the argument convention for packs into the PogsAttr and
out of the RefPackType.  The rationale here is that we'll eventually
move to a user defined struct as part of the formal type system, and
it won't need/want to have the arg convention as part of its type,
so we won't be able to recover it for that from the type checker.

Move in into PogsAttr which is where we put other argument related
metadata needed by the parser. While here, flatten the array of
packIndices into a single integer since an argument list can have at
most one pack.  We should do the same for VariadicIndices when
we get to a more flexible ParameterizedType representation.

MODULAR_ORIG_COMMIT_REV_ID: 2d75cf2960fd1403619ed7714639373a01a70713