[mojo-lang] Replace `lit.ref.pack.size` with `variadic_size` attr. NFC (#34640)
This removes the `lit.ref.pack.size` op that Ia added (still unused in
the tree) and replaces it with a ParamOperatorAttr that returns the
size of a !kgen.variadic list.  The type list for a pack is a variadic,
so we can get it at comptime and don't need anything special for packs.

This complements the existing variadic_get parameterattr.

MODULAR_ORIG_COMMIT_REV_ID: f94d597e9806316c75821582e962ab80cadbc911