[mojo-lang][KGEN] teach `*.pack.get` ops to infer a result type. (#34659)
We have a fully parametric type system, so we can infer a
parametric type from the results of these operations instead of
requiring that they be specified manually.  This sorts itself all
out post-elaboration.

This allows removing the T parameter to Tuple.get/ListLiteral.get
but I didn't do that, I'll file an issue.

modular-orig-commit: fb8d23be1d3a737ecf75fb6951aea473e25ae8c4