[Cache] Unbind runtime from Cache creation (#34498)
Up until now, the cache has required a Runtime for any creation and is tightly
bound with this. This is quite awkward, especially given that nearly no users
of the cache actually require asynchronous execution.

This change does a few things that are heavily interconnected:

* Add explicit Sync versions of cache methods and make them directly callable
  from the other callsites. In many cases, these are simply cheaper and more
  convenient (e.g. if you are executing from a pass manager thread, there is zero
  reason to attempt to schedule on a runtime thread).
* Move the Runtime requirement to the actual asychronous calls, instead of
  binding to the cache directly. This is tied in to a refactor of the blob
  cache calls that ties the behavior together correctly.
* Drop the "RuntimeAndCache" object, as this is no longer needed. The cache
  and runtime live on their own; cache calls which are asynchronous simply
  require a runtime to be passed in.

Why do this? The early binding on the runtime creates awkward requirements in
other parts of the code base. Lifting this means that we can plumb the runtime
with more ease elsewhere.

MODULAR_ORIG_COMMIT_REV_ID: 013f72f7207bd333f13b7532e9ac5e756e74b203