[mojo-stdlib] Change `Optional[T].or_else()` to return a `T` (#34683)
Change the return type to be a `T` instead of `Optional[T]`. It's not a
good API design to wrap the default value inside an optional if it's
caller-provided, since it *forces* the caller to call `.value()` on it. The caller always
knows it's engaged since they provided the value after all. If they want the
default value to be an `Optional`, they can just specify that as their
`default_value`. Do not wrap it for them always. 
It's slow and not good user ergonomics.

Fixes [Internal Link]

modular-orig-commit: 03bbc710f7adc16830a1c67b88b3c11125bb6dac