[Support] Fix DTypeSwitch with optional results (#34755)
Change DTypeSwitch to explicitly load the callable into the correct result type.
Only after it is loaded, move it into the optional.

Otherwise, if the result type is an optional, and std::nullopt is returned, it would set the outer optional.
This was happening here: [Internal link]

The otherwise returning std::nullopt was setting the outer optional for the DTypeSwitch as a whole.
As such, when hit, it would result in:
```
mt: /__w/modular/modular/Support/include/Support/ML/DType.h:468:
         ResultType M::DTypeSwitch<std::optional<llvm::SmallVector<float,
         12>>>::operator optional() [ResultType =
         std::optional<llvm::SmallVector<float, 12>>, ParamPtrTypes = <>]:
         Assertion `result && "Fell off the end of a DTypeSwitch"' failed.
```

This was breaking StarCoder in Pytorch.

MODULAR_ORIG_COMMIT_REV_ID: d2133360e7672e6d95db83bce7f2468d505ea72d