[Kernels][GPU] Add missing __copy_capture and re-enable softmax test (#34774)
This test was accidentally disabled for a long time, and in the meantime
the __copy_capture decorator was added to capture arguments by value.
The test was capturing the NDBuffer by reference which meant that the
captured buffer passed to the GPU kernel was actually a pointer to host
memory.

Fixes [Internal link]

GPU_TEST_ORIG_REV_ID: 5c0ca7e694e48bb5d20dc76a348ca986a3c56eff