[mojo-lang] Expose parameter printing on ASTType. NFC (#34947)
This just exposes a static method out of ASTType as a static method
so it can be used by mangling logic.  This whole thing could be
refactored to not duplicate stuff in mangling, but this is a logical
step forward.

MODULAR_ORIG_COMMIT_REV_ID: 9d1921876572caa4fea5353b3be1935f0dde171c