[mojo-lang] Make `VariadicPack` part of the function ABI. (#34950)
This teaches the caller to form a `VariadicPack` instance on the caller
side of a function with a pack, and pass it into the callee. This allows
proper memory management of owned arguments, and makes it so users of
variadic packs get nice APIs.  Homogenous variadics should follow the
same model.

This has two limitations currently:
 1) This only applies to owned pack arguments.
 2) This only works with AnyType bounds, not trait bounds.

The first will be straight-forward to fix, but will require untangling
stdlib expectations that packs are formed with `!kgen.pack` directly,
fixing Tuple etc.

The second will be a bit tricker, because we don't have something like
an "AnyMetaType" that allows expressing "packs of Stringable"s in the
type system.  This will require some exploration.

MODULAR_ORIG_COMMIT_REV_ID: 5d6914e65bfabd95c46d48a239c0165ffcb45d7b