[Kernels] add initial version of CPU flash attention (#34945)
This is an initial version of CPU flash attention that is not currently
wired into the system. There's still work to do to clean this up, add
tests, etc, but this gets something in to get things moving.

For a model like BERT, the kernel supports input lambda for K/V/mask so
the slice+transpose from the model is completely fused into the
attention op (Q needs a little more work). The kernel is a tiny bit
faster, but the big win is from avoiding the intermediate tensor copies.

For a model like whisper-encoder-onnx, the intermediate attention score
buffer is massive (1500x1500). Here flash attention does a great job of
not producing these intermediate buffers and pipelining the cache usage.
On a c6i.4xlarge instance, I'm seeing QPS numbers go from 1.1 to 1.4.

Whisper does not yet take full advantage of the slice+transpose fusion
due to some extra reshapes around the transposes that make the graph
compiler produce unneeded intermediates. I'll work on cleaning up the
MOGG code to invoke the kernel and add a test flag to enable that kernel
so that others can investigate this.

NN_ORIG_REV_ID: d5102377b2dc022d37a7569d3d6d7a7cb5ee42a9