[mojo-stdlib] Remove `StaticIntTuple` ctor from `DimList` (#34948)
Remove the ability to construct a `StaticIntTuple` from a `DimList`.  It is only
used by kernels, so provide a free function to achieve the same functionality.
By doing this, it allows us to move `Kernels/mojo/stdlib/list.mojo` to not be
part of the open source stdlib.  `index.mojo` will remain since open source
stdlib depends on `StaticIntTuple` and friends, but never used the `DimList`
constructor.

modular-orig-commit: 0990dbca33fc6b646dfa5f070a552bb169637cde