[mojo-driver] Assert on unused CLI arguments (#35084)
Closes #16500.

It's a Modular employee programming error to define a `mojo` command
line interface that takes an argument, such as `--target`, but that
doesn't actually "claim" that argument by checking its value in some
way. This patch adds a function to check all arguments for any unclaimed
ones, and assert if any are found. It also adds a call to that assert
function to all of the `mojo` commands. This may prevent a bad command
line argument from being released to the public.

It could be worthwhile, in a follow-up patch, to move all argument
parsing earlier in these commands. For example:

1. `mojo build` completely parses the input Mojo file before it ever
   checks that it has been passed a valid `-o` argument; it should exit
   with an error sooner if the user has passed in
   `-o /path/to/dir/that/does/not/exist/a.out`.
2. `-D` defines are parsed when the Mojo parser is invoked; in theory
   several validation errors could be returned to the user sooner.

This patch does move any argument parsing earlier, it just adds the
assertions.

MODULAR_ORIG_COMMIT_REV_ID: 5fba2d1154fe097158029305de3ace54780e9f0f