[Support] Add generic context to MDialect (#35126)
Certain passes require the use of an LLCL runtime. Unfortunately, there
is far too much implicit plumbing and code generation around the passes
to reliability plumb the runtime through to where it is needed
explicitly.

In general, the solution to such a problem might be to hang the Runtime
off the MLIRContext or other such state that is available to passes.
Unfortunately, there is no good general purpose way to do this. Instead,
the recommendation is to invent a synthetic dialect, and hang the
runtime state off of this. The MLIRContext serves as a holder of the
registry that allows access to this dialect, which contains the runtime.

This commit adds the MDialect, which holds a reference to a generic
context. This context can be used to hold the runtime and other things
over time, and replace the explicit runtime plumbing in LLVM passes.

MODULAR_ORIG_COMMIT_REV_ID: 842e949ddba4ccd2c1a677b068ace3aafc07ae23