[Support][DebugInfo] Sink kill Debug Values to the end of its source line (#34272)
Following this discussion:
[Internal link]
it was decided that we should sink kill Debug ValueOps so that they are
always the last instructions corresponding to its source line. This
bridges a semantic difference between our DebugInfo ValueOps & LLVM's
DbgValueOps:

DebugInfo dialect: ValueOp defines the value a variable takes on at that
point in the program. The position of the ValueOp in the IR, as well as
the source location of the ValueOp both have significance.
LLVM: DbgValueOp defines the value a variable takes on at the location
of the previous concrete instruction. The source location of the ValueOp
is ignored.

When we sink kill DebugValues to the end of its source line, we ensure
that it never gets attached to instructions from the previous line
(unless all operations for that line is optimized out, in which case
it's impossible to break on that line anyway). This trades off
instruction-level accuracy for line-level accuracy, which is decided to
be more important, since there are many more places already without
instruction-level accuracy.

MODULAR_ORIG_COMMIT_REV_ID: 6421a2071cd9590bf1d153ef34db4ff7c4437c31