[LLCL] Change runtime context structure (#35248)
Instead of the runtime *being* the global context, it now refers to a
shared context. This will allow us to make the context the higher-level
object, and the runtime will be a member of the context. For
compatibility during such a transition, we simply create a unique
context during runtime creation, and the semantics remain identical to
the current semantics (with the exception that it has pointer
indirection and will be inherited in the nested case).

MODULAR_ORIG_COMMIT_REV_ID: 84ae0dc177fb8b3ba55b35b3a828b12452d2602d