[mojo-stdlib] cleanup: Combine _optional into collections.optional as OptionalReg (#35220)
This is an alternative approach, as [discussed
[Internal link]

> Why not just put this register-passable Optional definition in
stdlib/collections/optional.mojo named as e.g. OptionalReg?

> Granted, it wouldn't have any other use in the stdlib itself, but to
me it seems consistent with the purpose of the standard library that we
provide useful vocabulary types, even in the presense of warty language
issues like the AnyType/AnyRegType distinction. The stdlib should strive
to be as useful as possible, even if there are things that exist only
because of current language limitations. As long as we're not
guaranteeing backwards compatibility, I think the bar to include a
useful type in the stdlib should be relatively low.

* Also combined the two test_optional.mojo files

Fixes #35118

BUFFER_ORIG_REV_ID: 2336831f6bbebc10f110222e9f4b62790c3f4415