[KGEN][DebugInfo] Lower kgen signature type into DI (#35326)
Fixes #35292.

Lower kgen.signature type into DI subroutine during debug type
conversion. Briefly discussed with Walter about the level of fidelity
for this. There's a lot more that can be done to provide more details
about the function type (e.g. all kinds of conventions), but that's not
going to provide much value today for the debugger (& I'm exploring some
ideas for a more general solution instead of point solutions). So for
now we can keep things simple with a straight-forward conversion into a
function pointer (note that it has to be a subroutine pointer, not a raw
subroutine, as that's how llvm represents it).

MODULAR_ORIG_COMMIT_REV_ID: 095e424890f19f2f4284a2d29d12160bc52e65a0