[mojo-tooling] Add a new magic `%#` to hide code in documentation (#35549)
This magic is used to hide lines of code within a code block when
generating documentation. This is useful when writing doc tests, where
an example may include portions specific for testability, that shouldn't
be showed in the generated documentation. For example:

```mojo
var value = 5
%# print(value)
```

will generate documentation of the form:

```mojo
var value = 5
```

modular-orig-commit: 3474d474aec8a441d38170fb7ce0a408bf363691