[mojo-lang] Slightly reduce the power of the `*_` operator in param lists. (#35605)
The `*_` syntax is sugar for some number of `_` expressions in a
parameter list.  Per discussion with Laszlo and others, we would
like to extend it to support keyword params as well as positional
ones, and the imagined use-cases for it are all served by it being
at the end of the parameter list.

As such, simplify the implementation by requiring it to be at the
end of the parameter list.

modular-orig-commit: 38eb864d42f6ffda71a05a235ad31c3d3fdf8a99