[mojo-lang] Revise the implementation of `_` and `*_` in param lists. (#35607)
This changes the implementation of `_` and `*_` to be magic syntax
within parameter lists instead of general expressions.  This means
that we correctly reject things like `var x = _` instead of accepting
it and failing to lower at code generation time.

The idea here is that `_` and `*_` are magic syntax that is
syntactically
allowed within a parameter list, not something that is general purpose.

By constraining it to that context, we make it so general `_` is the
"unloadable" discard pattern.  This also makes it clear that unary star
is the "unpack operation" which may or may of may not be something we
can handle generically in the future, but at least we haven't prescribed
some semantics to it that can't be supported generally.

This fixes #34362

MODULAR_ORIG_COMMIT_REV_ID: a54e2e5d4d23b5672faeda7b8bee666a3ea9e667