[Mojo] Emit specific error message when user calls or subscripts a module (#35415)
It is common to accidentally call or subscript a module instead of a
declaration within the module. This patch improves the error message for
this case, and also attempts to suggest a fix to the user when the
imported module contains a declaration with the same name.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 552b42954398adbcf12576a6077184d13adb15b1