[KGEN] Add Init::createContext support for `mojo` tool (#35365)
This switches the existing runtime create over to using the Init
package, which standardizes creation of the telemetry context,
runtime and entitlements.

For `build` and `run`, this is also registered in the `MDialect`, which
will be used to eliminate fragile runtime plumbing in the passes.

MODULAR_ORIG_COMMIT_REV_ID: a63fc59452bc2c9db0dc7994eb2f1a7faa4366d8