[KGEN] Emplace interpreter blob anyways if key is present (#35741)
I am not sure how this situation materializes, but there is apparently a
race condition where the interpreter attempts to create a blob where its
key is present but the value is not. In this case, set the value anyways
so that it can not be null.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: d3190624ce65c40a67a9667f94c33610003cf093