[mojo] Remove the external link functionality (#35972)
This is all of the kgen.link, kgen.external.func, etc. functionality
now that none of the package compilation flow uses it. Half of
this was added long ago as part of a hypothetical use case of
linking in external C++ libraries (which was never used/never worked
completely).

When we have a more concrete plan for external dependencies/linking,
we can revive whichever parts of this are necessary (and with a cleaner
integration into the compiler).

MODULAR_ORIG_COMMIT_REV_ID: 5b2211dea19f2eed841459319c6ead8406a6f123