[KGEN] Add Init::createContext support to `mojo-lsp-server` (#35367)
This standardizes the creation of global singletons.

Unfortunately, the MojoServer relied on an ability to **delete the
runtime** in order to shutdown safely. We could either plumb the
runtime or a separate context through the code here, or simply
add some tracking regarding the number of active tasks and
simply enable joining on this. Since this logic is fairly simple,
this is the route that was chosen.

MODULAR_ORIG_COMMIT_REV_ID: f9ed92b70c87b612b1d23e54a3e7ce2403b35d4f