[mojo] Add `mojo build-project` command in preview (#35836)
Add a `mojo build-project` subcommand that will, in a forthcoming
commit, interact with `mojo-build-server`. The command and its
documentation is only available to Modular employees who configure CMake
with `-DMODULAR_RELEASE_PACKAGE_BUILD=OFF`.

This is part of the Mojo project manifest & build tool work. For more
information, see [our roadmap]. A design doc on the build server itself
is forthcoming, and will be populated [here].

Gating the build in this way means that there has to be a default (the
way Modulers build when building locally), but also some alternative to
the default that is run in CI. This patch chooses to have
`build-project` be included in the default (everyone builds
`mojo build-project`), but to set it to `OFF` for release builds
(external users should not see `mojo build-project` show up in online
documentation, nor should they be able to run `mojo build-project`).
This was my choice because:

1. `mojo build-project` is in active development and will soon be dogfed
   internally. Any errors when building it should appear during local
   development (and get reverted); these only appearing in release
   workflows would be very annoying and become release-blockers.
2. It's not the end of the world if `mojo build-project` appears
   externally because we forget to configure it as `OFF` someplace. We
   have publicly announced that we're developing this; see
   [the community proposal].

Also, another design decision here was to make this an entirely new
command, as opposed to building this as part of `mojo build`. I chose a
new command because of the configuration issues above; it's simpler, in
my opinion, to conditionally include an entirely new command, as opposed
to conditionally including several lines of C++ inside of an existing
source file.

[our roadmap]: [Internal link]
[here]: [Internal link]
[the community proposal]: [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 9bfe5f3f3d87683683386b5d1ed1019491686ea6