[mojo] JSON transport for build server (#35911)
Add a C++ library for `mojo-build-server`, called `MojoBuild`. This is
part of the Mojo project manifest & build tool work. For more
information, see [our roadmap]. A design doc on the build server itself
is forthcoming, and will be populated [here].

The `MojoBuild` library will define the core logic for the build
server, as C++ routines that can be called from within a Mojo
executable. The idea is that, as we expand what can be done in Mojo,
we'll pull more logic out of C++ `MojoBuild` and into Mojo
`mojo-build-server`. For now, `MojoBuild` define a `mojoBuildServerMain`
function that waits to decode JSON-RPC messages.

No message handler is implemented yet, so this function can only ever
wait or error, for now. A forthcoming commit will add message handling,
and `mojo build-project` will be modified to start the server and send
it messages.

[our roadmap]: [Internal link]
[here]: [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: fd79aeff112c602b664e28f51a53945a9a7e11ab