[CI] Run standard library tests
Currently, `examples.yml` only runs the examples, which was convenient for
launching today.  Now that we've shipped the standard library though, augment
the workflow file to also run the standard library tests.

While we're here, rename the name of the file, step, and name of the job to
better reflect reality that this file is used to run the standard library tests
and the examples now.