[Support] Simplify HTTP auth and loading of certificates (#35689)
If the certificates are provided by an environment variable, then they
cannot be loaded from the file system. This change drops this behavior
in favor of using the in-memory certificates.

Further, this simplifies some of the auth paths. Because the context is
a global singleton, calls to `noAuthNeeded` were essentially clearing
existing keys destructively. These are no longer necessary, and the
client can function either with or without keys, which are provided when
the initial context is created.

MODULAR_ORIG_COMMIT_REV_ID: 99976003ab94fa8ef5834aa58b810ffd3fcc29d3