[All] Support for Init::createContext (#35370)
This switches the installer and associated code to use the global
context. The general logic is kept the same.

In order to support these changes, some additional logic has been added
to Settings and the HTTPContext. Namely:
* The HTTPContext supports a CAInfo attribute, which allows us to set
the SSL certificate for testing.
* The settings correctly fall back to the configuration when a key is
not there. This required adding appropriate thread-safety in the form of
a map around the lock.
* MSupportInit drops the LLCLRuntimeGlobals requirement, as this causes
a dependency on the static library. This will generally be needed by
individual tools, but should not be added at the `modular_library`
level.

MODULAR_ORIG_COMMIT_REV_ID: cffd051172677dfe634fdb07b5a7b14c1ca7db14