[driver-tblgen] Add support for trailing input arguments (#36152)
This adds a new bit to UsageInfo, `inNameHasArgs` that indicates if
the input has arguments, in which case a trailing `[inputName-arguments...]`
is generated in the usage. This is useful for commands that have
different sets of arguments for the input (such as `mojo run`).

MODULAR_ORIG_COMMIT_REV_ID: c34311ceb868c6695a22d73c4020e1c97aeee6a6