[driver-tblgen] Add help text for AliasArgs (#36177)
AliasArgs are used to specify arguments of an option that an alias
represents, for example in `mojo` the `-g` option aliases
`--debug-level full`. This PR ensures that the arguments that an alias
sets are displayed in the help text. For `-g`, this is now `-g (LEVEL=full)`.

MODULAR_ORIG_COMMIT_REV_ID: 9d169a672883202f4e098164981bf0759b5941b4