[Support] Add `remove` operation to the context (#36219)
This shouldn't generally be used, but is a nice escape hatch. For
example, currently the `LLCL::Runtime` can be removed from the context
at the end in order to force a flush of the allocated chains or writing
of the trace profile. We need a better mechanism for doing these things,
but this provides a stop gap until we have such things.

MODULAR_ORIG_COMMIT_REV_ID: 707218668b880c534bfdf1c7569413b5621af063