[MojoBuild] Find `build_server_path` in config (#36237)
Use the `modular.cfg` to find the path to the `mojo-build-server`,
instead of relying on its installation location. This is more in line
with all other tools in the Mojo SDK, and works reliably in all
environments.

This is part of the Mojo project manifest & build tool work. For more
information, see [our roadmap]. A design doc on the build server itself
is forthcoming, and will be populated [here].

Resolves feedback left on #36106.

[our roadmap]:
[Internal link]
[here]:
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 4b373660064d042fef83cde7eb45f754d3c5d1d1