[mojo-lang] Revise implicit parameters for traits. (#36324)
This changes the internal implementation of traits from having to
implicit parameters to having a single parameter T of the same
type as the trait.  This is an easier-for-my-brain to understand
model than taking a metatype and the trait itself.

Doing this requires some non-trivial attribute rewriting gymnastics,
(both in overload checking and in requirement inheritance) but leads
to a simpler to work with model.

This fixes #25492

MODULAR_ORIG_COMMIT_REV_ID: 20af867689393e387fa6423f15797d4422c5656f