[bazel] Move lit cfg generation to a custom rule (#36419)
When using a genrule with `tools`, anything referenced in the tools
attribute is compiled for the exec configuration. In our case we really
want them to be built for the target configuration since these
executables are executed during test execution. Unfortunately genrule
doesn't provide a standard data attribute to allow us to do that, so
instead we have to use a custom rule. The downside of having split
configurations for now is you have to rebuild everything twice. In the
future with remote execution this could be a bigger issue since the
binaries produced as a dependency of lit tests need to run on the target
machine and not the machine orchestrating the remote build.

MODULAR_ORIG_COMMIT_REV_ID: dc3125f429736e925e7fb56c506d853aaab30982