[KGEN] Fix crash when parsing `0__` (#35899)
There was an assumption that if a literal starts with `0` and is at
least 3 characters long, it has to be because it started with `0b`,
`0x`, etc., and hence was safe to strip off 2 characters. When you do
this with `0__`, you end up with `_`, which after filtering underscores,
is an empty string, which can't be parsed as an integer, causing an
assertion error. Instead, the first two characters only if it's one of
the prefixes we expect.

Fixes modularml/mojo#1913.

modular-orig-commit: 75cce9ad8716af3f950a5e36fa855a343f65ba16