Allow `my_optional is None` and `my_optional is not None` (#2082)
Just as the title says, it's a very common idiom in python, is even recommended by PEP8: https://peps.python.org/pep-0008/#programming-recommendations.

Comparisons to singletons like `None` should always be done with is or is not, never the equality operators.
Also, beware of writing if x when you really mean if x is not None – e.g. when testing whether a variable or argument that defaults to None was set to some other value. The other value might have a type (such as a container) that could be false in a boolean context!

While the warning does not apply to Mojo, Python users are quite used to it and it's a small quality of life improvement which is inexpensive to support.

Signed-off-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>