[mojo-lang] Fix `VariadicPack` with no arguments in dynamic case. (#36676)
We need to bind the lifetime of a variadic pack with no arguments
to something valid instead of leaving it the declared type.
Refactor the three callers of `emitVariadicPackConstructor` a bit
to share the logic to make sure this isn't dropped on one of them,
fixing a crash.

MODULAR_ORIG_COMMIT_REV_ID: e301d71c859cad7d17549900af24fb1e2b76529b