[KGEN] Ensure that mblack is copied when the grammar file is changed (#36693)
The grammar file is the most common thing that changes, so at minimum we
should make sure that we copy mblack if that changes. In theory we
should copy whenever any part of the directory changes, but that
requires a knowledge of cmake that I do not possess.

MODULAR_ORIG_COMMIT_REV_ID: 729408c3e0240256350d7efa09923124dbdb3dc6