[mojo-lang][CheckLifetimes] Fix obscure crash Stef ran into (#36773)
This fixes destructor insertion to notice if a reference is to a
non-default address space.  We cannot destroy these things, because
the `self` of a `__del__` method is always in the default addrspace.

Instead of crashing the compiler, emit an error.

MODULAR_ORIG_COMMIT_REV_ID: 2de6d805b0a3d0008376d7e715e0a1841840b4da