[mojo-tooling][Docs] Overhaul handling of variadicness in doc generation (#36735)
Instead of trying to reason about variadicness based on some hacky
string checking and manipulation, we can propagate it into the
`{Parameter,Argument}DeclView` structs, similarly to how we propagate
passing kind. This allows fixing the incorrect placement of `*` in the
emitted signature string. The patch also goes a step further, and
changes how we print identifiers and types in the generated JSON:
instead of prepeding them to the type, we prepend them to the name, so
that it is consistent with how it looks in the signature. The only case
when the type needs a `*` prepended is in pack arguments, since the
variadic type is _sort of_ unpacked there.

Partially addresses [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: d851fb299d3fef6a009cab19a6f4f250cf76f7d7