[bazel] Fix crash reporting lit tests (#36925)
Some lit tests in Support/ depend on finding the
modular-crashpad-handler, which right now in cmake comes from the
modular derived path. We don't attempt to emulate that in bazel so
instead we can forward the env var with the bazel path instead. This
requires a slight test change to unset this in the case it's testing
behavior of overriding this in the config, since the env takes
precedence.

MODULAR_ORIG_COMMIT_REV_ID: 3c5764bca62c6823837cef7d521d5150f9d59f09