[KGEN][DebugInfo] Strip debug scope if no DIBuilder (#36863)
Fixes #36717.

When setting a scoped location on a function, if DIBuilder does not
exist (i.e. building in no-debug mode), the function needs to be checked
for an existing scope and have it removed. This solves the case where
functions need to be synthesized on package import for a no-debug
compile. Otherwise, the struct comes with a debug scope (DIFile) from
the package, but there is no new DISubprogram scope to assign to the
function, which causes the function to inherit the file scope, which is
illegal.

This is distinct from [the other
[Internal link] where
synthesized functions are built with debug scopes under debug mode.

MODULAR_ORIG_COMMIT_REV_ID: 8d048b88bb65d588583420eb3e4fb4be3de6f6e8